refatoração

+++ devices/robotCar/src/main/java/com/eliascoelho911/homeapp/robotcar/domain/model/Speed.kt
@@ -1,3 +1,3 @@
 package com.eliascoelho911.homeapp.robotcar.domain.model
 
-internal typealias Speed = Float
+internal typealias Speed = Int
